chore: Split timeline widget files
